Welcome Plumbio
Make the appropriate changes related to moving from Guardio to Plumbio.

While Guardio received data and simply analyzed it — using its own
guards or custom user-defined ones — to tell if it was valid or not,
Plumbio is that and much more.

Plumbio still has the data-analyzing aspects, but it can also be used to
act on the data as well, with the possibility of pipes made to modify it
— or not, depending on how the stream flows through each system.

With Guardio variables, you had the data to be analyzed and a separate
object, called payload, intended to be passed down from guard to guard.

In Plumbio, these two variables are merged, and all data is passed down
from pipe to pipe so that it can either be analyzed or modified at any
moment.

This new variable is called "stream", like a stream of data flowing
through the pipes.